requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cloudinary==1.36.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.22
django-allauth==0.53.1
django-bootstrap-datepicker-plus==5.0.4
django-countries==7.2.1
django-crispy-forms==1.14.0
django-tinymce==3.6.1
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==7.1.0
whitenoise==6.6.0
